file refactorization completed in assets>js
