+++ server.js
@@ -1,6 +1,6 @@
 const Discord = require("discord.js");
 const client = new Discord.Client();
-const presence = require("./src/assets/js/misc/presence");
+const presence = require("./src/assets/js/client/status/presence");
 const david = require("./src/responses/david");
 require("dotenv").config();
