+++ src/assets/js/client/status/presence.js
