+++ src/assets/js/message/content/contentFlags.js
