+++ src/responses/david.js
@@ -1,4 +1,4 @@
-const messageFlags = require("../assets/js/misc/messageContentFlags");
+const messageFlags = require("../assets/js/message/content/contentFlags");
 
 module.exports = (message) => {
   const [command, flag] = messageFlags(message);
