+++ src/responses/image_generator/singleWorded.js
@@ -1,0 +1,3 @@
+module.exports = (() => {
+    
+})();
